test: move test files to a dedicated sub-folder
The test root folder already contained configuration scripts and subfolders for shared utilities and
BPMN diagrams. The test files were hidden in all this.
A new "spec" subfolder now stores all test files for greater clarity (a central location for all
test files).
